MultiAgentGraph/knowledge_graph_builder/requirements.txt
# This project uses setup.py for dependency management
# To install the package and its dependencies, use:
#
# For development (includes all dev tools):
# pip install -e .[dev]
#
# For production (core dependencies only):
# pip install -e .
#
# For documentation:
# pip install -e .[docs]
#
# For testing:
# pip install -e .[test]
#
# See setup.py for detailed dependency information.

# Core dependencies (for quick reference)
streamlit>=1.32.0
requests>=2.31.0
google-search-results>=2.4.2
wikipedia>=1.4.0
python-dotenv>=1.0.0
graphviz>=0.20.1
openai>=1.12.0
typing-extensions>=4.9.0

# Development Tools (optional)
# ---------------------------
# Testing
pytest>=8.0.0
pytest-cov>=4.1.0  # For test coverage reports

# Code Quality
black>=24.1.0  # Code formatting
flake8>=7.0.0  # Linting
isort>=5.13.0  # Import sorting

# Documentation
sphinx>=7.2.0  # For generating documentation
sphinx-rtd-theme>=2.0.0  # Documentation theme

# Note: Install development tools with:
# pip install -r requirements.txt[dev]
# Or install only core dependencies with:
# pip install -r requirements.txt

